♻️ refactor: Move utility functions to util folder
- Moved utility functions to a dedicated 'util' folder for better organization
