+++ src/components/Media/MediaSection/GridView/index.jsx
@@ -3,30 +3,15 @@ import { useContext, useEffect, useState } from 'react';
 import ConfirmModal from '@components/common/Modal';
 
 import { SubscribeContext } from '@/context/SubscribeContext';
+import {
+  getPressDataWithSubscription,
+  filterSubscribedPressData,
+  getSlicedData,
+} from '@/utils/parse';
 
 import GridView from './GridView';
 import Carousel from '../Carousel';
 import useCarousel from '../Carousel/useCarousel';
-
-//유틸
-
-// 구독 상태를 포함한 데이터를 준비
-const getPressDataWithSubscription = (pressList, subscribedPress) => {
-  return pressList.map((press) => ({
-    ...press,
-    isSubscribed: subscribedPress.includes(press.pid),
-  }));
-};
-
-//필터 함수
-const filterSubscribedPressData = (pressDataWithSubscription) => {
-  return pressDataWithSubscription.filter((press) => press.isSubscribed);
-};
-
-//데이터 자르는 함수
-const getSlicedData = (data, startIndex, endIndex) => {
-  return data.slice(startIndex, endIndex);
-};
 
 const GridViewContainer = ({ data: pressList, activeTab }) => {
   const { currentPage, goNext, goPrev, reset } = useCarousel();
